Avoid copying instruction data when recording instructions
