+++ svm/src/transaction_processor.rs
@@ -23,11 +23,6 @@ use {
     solana_account::{state_traits::StateMut, AccountSharedData, ReadableAccount, PROGRAM_OWNERS},
     solana_clock::{Epoch, Slot},
     solana_hash::Hash,
-    solana_instruction::TRANSACTION_LEVEL_STACK_HEIGHT,
-    solana_message::{
-        compiled_instruction::CompiledInstruction,
-        inner_instruction::{InnerInstruction, InnerInstructionsList},
-    },
     solana_nonce::{
         state::{DurableNonce, State as NonceState},
         versions::Versions as NonceVersions,
@@ -935,20 +930,15 @@ impl<FG: ForkGraph> TransactionBatchProcessor<FG> {
                     .ok()
             });
 
-        let inner_instructions = if config.recording_config.enable_cpi_recording {
-            Some(Self::inner_instructions_list_from_instruction_trace(
-                &transaction_context,
-            ))
-        } else {
-            None
-        };
+        let (execution_record, inner_instructions) = transaction_context
+            .deconstruct_into_record(config.recording_config.enable_cpi_recording);
 
         let ExecutionRecord {
             accounts,
             return_data,
             touched_account_count,
             accounts_resize_delta: accounts_data_len_delta,
-        } = transaction_context.into();
+        } = execution_record;
 
         if status.is_ok()
             && transaction_accounts_lamports_sum(&accounts)
@@ -983,54 +973,6 @@ impl<FG: ForkGraph> TransactionBatchProcessor<FG> {
             loaded_transaction,
             programs_modified_by_tx: program_cache_for_tx_batch.drain_modified_entries(),
         }
-    }
-
-    /// Extract the InnerInstructionsList from a TransactionContext
-    fn inner_instructions_list_from_instruction_trace(
-        transaction_context: &TransactionContext,
-    ) -> InnerInstructionsList {
-        debug_assert!(transaction_context
-            .get_instruction_context_at_index_in_trace(0)
-            .map(|instruction_context| instruction_context.get_stack_height()
-                == TRANSACTION_LEVEL_STACK_HEIGHT)
-            .unwrap_or(true));
-        let mut outer_instructions = Vec::new();
-        for index_in_trace in 0..transaction_context.get_instruction_trace_length() {
-            if let Ok(instruction_context) =
-                transaction_context.get_instruction_context_at_index_in_trace(index_in_trace)
-            {
-                let stack_height = instruction_context.get_stack_height();
-                if stack_height == TRANSACTION_LEVEL_STACK_HEIGHT {
-                    outer_instructions.push(Vec::new());
-                } else if let Some(inner_instructions) = outer_instructions.last_mut() {
-                    let stack_height = u8::try_from(stack_height).unwrap_or(u8::MAX);
-                    let instruction = CompiledInstruction::new_from_raw_parts(
-                        instruction_context
-                            .get_index_of_program_account_in_transaction()
-                            .unwrap_or_default() as u8,
-                        instruction_context.get_instruction_data().to_vec(),
-                        (0..instruction_context.get_number_of_instruction_accounts())
-                            .map(|instruction_account_index| {
-                                instruction_context
-                                    .get_index_of_instruction_account_in_transaction(
-                                        instruction_account_index,
-                                    )
-                                    .unwrap_or_default() as u8
-                            })
-                            .collect(),
-                    );
-                    inner_instructions.push(InnerInstruction {
-                        instruction,
-                        stack_height,
-                    });
-                } else {
-                    debug_assert!(false);
-                }
-            } else {
-                debug_assert!(false);
-            }
-        }
-        outer_instructions
     }
 
     pub fn fill_missing_sysvar_cache_entries<CB: TransactionProcessingCallback>(
@@ -1093,7 +1035,10 @@ mod tests {
         solana_fee_structure::FeeDetails,
         solana_hash::Hash,
         solana_keypair::Keypair,
-        solana_message::{LegacyMessage, Message, MessageHeader, SanitizedMessage},
+        solana_message::{
+            compiled_instruction::CompiledInstruction, LegacyMessage, Message, MessageHeader,
+            SanitizedMessage,
+        },
         solana_nonce as nonce,
         solana_program_runtime::{
             execution_budget::{
@@ -1106,7 +1051,6 @@ mod tests {
         solana_signature::Signature,
         solana_svm_callback::{AccountState, InvokeContextCallback},
         solana_transaction::{sanitized::SanitizedTransaction, Transaction},
-        solana_transaction_context::TransactionContext,
         solana_transaction_error::{TransactionError, TransactionError::DuplicateInstruction},
         std::collections::HashMap,
         test_case::test_case,
@@ -1242,60 +1186,6 @@ mod tests {
             check_results,
             &TransactionProcessingEnvironment::default(),
             &TransactionProcessingConfig::default(),
-        );
-    }
-
-    #[test]
-    fn test_inner_instructions_list_from_instruction_trace() {
-        let instruction_trace = [1, 2, 1, 1, 2, 3, 2];
-        let mut transaction_context = TransactionContext::new(
-            vec![(
-                Pubkey::new_unique(),
-                AccountSharedData::new(1, 1, &bpf_loader::ID),
-            )],
-            Rent::default(),
-            3,
-            instruction_trace.len(),
-        );
-        for (index_in_trace, stack_height) in instruction_trace.into_iter().enumerate() {
-            while stack_height <= transaction_context.get_instruction_stack_height() {
-                transaction_context.pop().unwrap();
-            }
-            if stack_height > transaction_context.get_instruction_stack_height() {
-                transaction_context
-                    .configure_next_instruction_for_tests(0, vec![], &[index_in_trace as u8])
-                    .unwrap();
-                transaction_context.push().unwrap();
-            }
-        }
-        let inner_instructions =
-            TransactionBatchProcessor::<TestForkGraph>::inner_instructions_list_from_instruction_trace(
-                &transaction_context,
-            );
-
-        assert_eq!(
-            inner_instructions,
-            vec![
-                vec![InnerInstruction {
-                    instruction: CompiledInstruction::new_from_raw_parts(0, vec![1], vec![]),
-                    stack_height: 2,
-                }],
-                vec![],
-                vec![
-                    InnerInstruction {
-                        instruction: CompiledInstruction::new_from_raw_parts(0, vec![4], vec![]),
-                        stack_height: 2,
-                    },
-                    InnerInstruction {
-                        instruction: CompiledInstruction::new_from_raw_parts(0, vec![5], vec![]),
-                        stack_height: 3,
-                    },
-                    InnerInstruction {
-                        instruction: CompiledInstruction::new_from_raw_parts(0, vec![6], vec![]),
-                        stack_height: 2,
-                    },
-                ]
-            ]
         );
     }
