+++ transaction-context/Cargo.toml
@@ -26,6 +26,7 @@ serde_derive = { workspace = true, optional = true }
 solana-account = { workspace = true }
 solana-instruction = { workspace = true, features = ["std"] }
 solana-instructions-sysvar = { workspace = true }
+solana-message = { workspace = true }
 solana-pubkey = { workspace = true }
 solana-sbpf = { workspace = true }
